switched to ruff and split dependencies, adjusting width of elements and navbar for better mobile view
