+++ budgetdb/templates/budgetdb/_navbar.html
@@ -1,5 +1,7 @@
+<!-- budgetdb/templates/budgetdb/_navbar.html -->
+
 {% load static %}
-<nav class="navbar navbar-expand-lg fixed-top bg-body-tertiary" >
+<nav class="navbar navbar-expand-lg sticky-top bg-body-tertiary" >
     <div class="container-fluid">
     <a class="navbar-brand" href="{% url 'budgetdb:home' %}">Budget</a>
     <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
