+++ budgetdb/templates/budgetdb/base.html
@@ -31,7 +31,7 @@
         </script>
     </head>
 
-    <body style="padding-top: 75px;">
+    <body >
         {% csrf_token %}
         
         {% block navtools %}
