+++ budgetdb/templates/budgetdb/dashboard.html
@@ -37,7 +37,7 @@
         </div>
         <div class="col">
             <div class="card">
-                <div class="card-body" style="width: 600px;height:400px;">
+                <div class="card-body">
                     <h5 class="card-title">This year</h5>
                     <div class="echart-container" 
                         id="chart4"
