+++ budgetdb/templates/budgetdb/transaction/base_transactions_list.html
@@ -19,12 +19,12 @@
 
     <div class="row align-items-center"> 
         {% if filter_type == 'account' %}
-            <div class="col-1 text-start">   
+            <div class="col-2 text-start">   
                 <button  type="button" title="Create another transaction for today" class="update-transaction btn btn-link btn-sm" data-form-url="{% url 'budgetdb:create_transaction_from_date_and_account' pk %}" >
                     <span class="material-symbols-outlined">add_circle</span>
                 </button>
             </div>
-            <div class="col-10 text-center">
+            <div class="col-8 text-center">
                 <h3>{{ account_name }} 
                 <a href="{% url 'budgetdb:year_report_account' pk year %}" class="btn btn-info btn-sm"  >
                     <span class="material-symbols-outlined">monitoring</span></a>
@@ -34,7 +34,7 @@
                 {% endif %}
                 </h3>
             </div>
-            <div class="col-1 text-end">
+            <div class="col-2 text-end">
                 <button  type="button" title="Create audit for today" class="update-transaction btn btn-link btn-sm" data-form-url="{% url 'budgetdb:create_audit_transaction_from_account' pk %}" >
                     <span class="material-symbols-outlined">add_circle</span>
                 </button>
