+++ requirements-dev.txt
@@ -1,0 +1,174 @@
+#
+# This file is autogenerated by pip-compile with Python 3.12
+# by the following command:
+#
+#    pip-compile requirements-dev.in
+#
+anyio==4.13.0
+    # via httpx
+asgiref==3.11.0
+    # via
+    #   -c requirements.txt
+    #   django
+astroid==4.0.4
+    # via
+    #   -r requirements-dev.in
+    #   pylint
+certifi==2026.5.20
+    # via
+    #   -c requirements.txt
+    #   httpcore
+    #   httpx
+    #   requests
+charset-normalizer==3.4.7
+    # via
+    #   -c requirements.txt
+    #   requests
+click==8.4.1
+    # via
+    #   -c requirements.txt
+    #   codecov-cli
+codecov-cli==0.1.13
+    # via -r requirements-dev.in
+coverage[toml]==7.14.1
+    # via
+    #   -r requirements-dev.in
+    #   pytest-cov
+crispy-bootstrap5==2026.3
+    # via -r requirements-dev.in
+dill==0.4.1
+    # via
+    #   -c requirements.txt
+    #   pylint
+django==6.0.5
+    # via
+    #   -c requirements.txt
+    #   crispy-bootstrap5
+    #   django-crispy-forms
+django-crispy-forms==2.6
+    # via
+    #   -c requirements.txt
+    #   crispy-bootstrap5
+h11==0.16.0
+    # via httpcore
+httpcore==1.0.9
+    # via httpx
+httpx==0.28.1
+    # via codecov-cli
+idna==3.18
+    # via
+    #   -c requirements.txt
+    #   anyio
+    #   httpx
+    #   requests
+ijson==3.5.0
+    # via
+    #   -c requirements.txt
+    #   codecov-cli
+iniconfig==2.3.0
+    # via
+    #   -r requirements-dev.in
+    #   pytest
+isort==8.0.1
+    # via
+    #   -r requirements-dev.in
+    #   pylint
+lxml==6.1.1
+    # via
+    #   -c requirements.txt
+    #   -r requirements-dev.in
+mccabe==0.7.0
+    # via
+    #   -r requirements-dev.in
+    #   pylint
+packaging==26.2
+    # via
+    #   -c requirements.txt
+    #   pytest
+platformdirs==4.10.0
+    # via
+    #   -r requirements-dev.in
+    #   pylint
+pluggy==1.6.0
+    # via
+    #   -r requirements-dev.in
+    #   pytest
+    #   pytest-cov
+pycodestyle==2.14.0
+    # via -r requirements-dev.in
+pycparser==3.0
+    # via
+    #   -c requirements.txt
+    #   -r requirements-dev.in
+pyee==13.0.1
+    # via -r requirements-dev.in
+pygments==2.20.0
+    # via
+    #   -r requirements-dev.in
+    #   pytest
+pylint==4.0.5
+    # via -r requirements-dev.in
+pypdfium2==5.9.0
+    # via
+    #   -c requirements.txt
+    #   -r requirements-dev.in
+pytest==9.0.3
+    # via
+    #   -r requirements-dev.in
+    #   codecov-cli
+    #   pytest-base-url
+    #   pytest-cov
+    #   pytest-django
+pytest-base-url==2.1.0
+    # via -r requirements-dev.in
+pytest-cov==7.1.0
+    # via
+    #   -r requirements-dev.in
+    #   codecov-cli
+pytest-django==4.12.0
+    # via -r requirements-dev.in
+pyyaml==6.0.3
+    # via
+    #   -c requirements.txt
+    #   codecov-cli
+    #   responses
+requests==2.34.2
+    # via
+    #   -c requirements.txt
+    #   codecov-cli
+    #   pytest-base-url
+    #   responses
+responses==0.26.1
+    # via -r requirements-dev.in
+ruff==0.15.16
+    # via -r requirements-dev.in
+smart-open==7.6.1
+    # via codecov-cli
+sqlparse==0.5.5
+    # via
+    #   -c requirements.txt
+    #   django
+test-results-parser==0.6.1
+    # via -r requirements-dev.in
+tomli==2.4.1
+    # via -r requirements-dev.in
+tomlkit==0.15.0
+    # via
+    #   -r requirements-dev.in
+    #   pylint
+tree-sitter==0.25.2
+    # via codecov-cli
+typing-extensions==4.15.0
+    # via
+    #   -c requirements.txt
+    #   anyio
+    #   pyee
+urllib3==2.7.0
+    # via
+    #   -c requirements.txt
+    #   requests
+    #   responses
+wrapt==2.2.1
+    # via
+    #   -c requirements.txt
+    #   smart-open
