+++ requirements.in
@@ -1,0 +1,48 @@
+APScheduler
+asgiref
+beautifulsoup4
+certifi
+cffi
+charset-normalizer
+click==8.4.1
+cryptography
+dill
+Django
+django-appconf
+django-auditlog
+django-bootstrap-modal-forms
+django-chartjs
+django-crispy-forms
+django-crum
+django-environ
+django-extensions
+django-filter
+django-select2
+django-tables2
+djangorestframework
+greenlet
+gunicorn
+holidays
+idna
+ijson
+mysqlclient
+ofxparse
+packaging
+pdfminer.six
+pdfplumber
+pillow
+python-dateutil
+python-slugify
+pytz
+PyYAML
+requests
+sentry-sdk
+six
+soupsieve
+sqlparse
+text-unidecode
+typing_extensions
+tzlocal
+urllib3
+whitenoise
+wrapt
