+++ requirements.txt
@@ -1,73 +1,162 @@
-APScheduler==3.11.2
+#
+# This file is autogenerated by pip-compile with Python 3.12
+# by the following command:
+#
+#    pip-compile requirements.in
+#
+apscheduler==3.11.2
+    # via -r requirements.in
 asgiref==3.11.0
-astroid==3.3.10
+    # via
+    #   -r requirements.in
+    #   django
 beautifulsoup4==4.14.3
-certifi==2026.1.4
+    # via
+    #   -r requirements.in
+    #   ofxparse
+certifi==2026.5.20
+    # via
+    #   -r requirements.in
+    #   requests
+    #   sentry-sdk
 cffi==2.0.0
-charset-normalizer==3.4.4
-click==8.2.1
-codecov-cli==11.2.6
-coverage==7.13.4
-crispy-bootstrap5==2023.10
-cryptography==47.0.0
-dill==0.4.0
-Django==6.0.5
+    # via
+    #   -r requirements.in
+    #   cryptography
+charset-normalizer==3.4.7
+    # via
+    #   -r requirements.in
+    #   pdfminer-six
+    #   requests
+click==8.4.1
+    # via -r requirements.in
+cryptography==48.0.0
+    # via
+    #   -r requirements.in
+    #   pdfminer-six
+dill==0.4.1
+    # via -r requirements.in
+django==6.0.5
+    # via
+    #   -r requirements.in
+    #   django-appconf
+    #   django-auditlog
+    #   django-bootstrap-modal-forms
+    #   django-crispy-forms
+    #   django-crum
+    #   django-extensions
+    #   django-filter
+    #   django-select2
+    #   django-tables2
+    #   djangorestframework
 django-appconf==1.2.0
+    # via
+    #   -r requirements.in
+    #   django-select2
 django-auditlog==3.4.1
+    # via -r requirements.in
 django-bootstrap-modal-forms==3.0.5
+    # via -r requirements.in
 django-chartjs==2.3.0
-django-crispy-forms==2.1
+    # via -r requirements.in
+django-crispy-forms==2.6
+    # via -r requirements.in
 django-crum==0.7.9
-django-environ==0.12.0
+    # via -r requirements.in
+django-environ==0.13.0
+    # via -r requirements.in
 django-extensions==4.1
-django-filter==25.1
+    # via -r requirements.in
+django-filter==25.2
+    # via -r requirements.in
 django-select2==8.4.8
-django-tables2==2.7.5
-djangorestframework==3.16.1
-greenlet==3.3.1
-gunicorn==23.0.0
-holidays==0.93
-idna==3.16
-ijson==3.4.0.post0
-iniconfig==2.3.0
-isort==6.0.1
-lxml==6.1.0
-mccabe==0.7.0
-mysqlclient==2.2.7
+    # via -r requirements.in
+django-tables2==3.0.0
+    # via -r requirements.in
+djangorestframework==3.17.1
+    # via -r requirements.in
+greenlet==3.5.1
+    # via -r requirements.in
+gunicorn==26.0.0
+    # via -r requirements.in
+holidays==0.98
+    # via -r requirements.in
+idna==3.18
+    # via
+    #   -r requirements.in
+    #   requests
+ijson==3.5.0
+    # via -r requirements.in
+lxml==6.1.1
+    # via ofxparse
+mysqlclient==2.2.8
+    # via -r requirements.in
 ofxparse==0.21
-packaging==25.0
-pdfminer.six==20251230
+    # via -r requirements.in
+packaging==26.2
+    # via
+    #   -r requirements.in
+    #   gunicorn
+pdfminer-six==20251230
+    # via
+    #   -r requirements.in
+    #   pdfplumber
 pdfplumber==0.11.9
+    # via -r requirements.in
 pillow==12.2.0
-pip-review==1.3.0
-platformdirs==4.3.8
-pluggy==1.6.0
-pycodestyle==2.14.0
+    # via
+    #   -r requirements.in
+    #   pdfplumber
 pycparser==3.0
-pyee==13.0.1
-Pygments==2.20.0
-pylint==3.3.7
-pypdfium2==5.5.0
-pytest==9.0.3
-pytest-base-url==2.1.0
-pytest-cov==7.0.0
-pytest-django==4.11.1
+    # via cffi
+pypdfium2==5.9.0
+    # via pdfplumber
 python-dateutil==2.9.0.post0
+    # via
+    #   -r requirements.in
+    #   django-auditlog
+    #   holidays
 python-slugify==8.0.4
-pytz==2025.2
-PyYAML==6.0.3
-requests==2.33.1
-responses==0.21.0
-sentry-sdk==2.52.0
+    # via -r requirements.in
+pytz==2026.2
+    # via -r requirements.in
+pyyaml==6.0.3
+    # via -r requirements.in
+requests==2.34.2
+    # via -r requirements.in
+sentry-sdk==2.61.1
+    # via -r requirements.in
 six==1.17.0
-soupsieve==2.8.3
+    # via
+    #   -r requirements.in
+    #   ofxparse
+    #   python-dateutil
+soupsieve==2.8.4
+    # via
+    #   -r requirements.in
+    #   beautifulsoup4
 sqlparse==0.5.5
-test_results_parser==0.6.1
+    # via
+    #   -r requirements.in
+    #   django
 text-unidecode==1.3
-tomli==2.2.1
-tomlkit==0.13.3
-typing_extensions==4.14.0
+    # via
+    #   -r requirements.in
+    #   python-slugify
+typing-extensions==4.15.0
+    # via
+    #   -r requirements.in
+    #   beautifulsoup4
 tzlocal==5.3.1
+    # via
+    #   -r requirements.in
+    #   apscheduler
 urllib3==2.7.0
-whitenoise==6.9.0
-wrapt==1.17.2
+    # via
+    #   -r requirements.in
+    #   requests
+    #   sentry-sdk
+whitenoise==6.12.0
+    # via -r requirements.in
+wrapt==2.2.1
+    # via -r requirements.in
